Makefile.in: Add link to library
When the library is installed a libfaifa.so.0 is installed but there is
no symbolic link created to it.

This can confuse some tools where they are looking for libfaifa when you
are trying to link.

Note: Ideally, libtool should be used and a .pc files (pkg-config)
should be created.

Fixes #5

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@devialet.com>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>